+++ src/main/java/com/example/solidconnection/community/controller/board/BoardController.java
@@ -1,14 +1,10 @@
 package com.example.solidconnection.community.controller.board;
 
-import com.example.solidconnection.community.service.board.BoardService;
-import com.example.solidconnection.community.dto.post.BoardFindPostResponse;
 import com.example.solidconnection.type.BoardCode;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 import java.util.ArrayList;
@@ -19,8 +15,6 @@ import java.util.List;
 @RequestMapping("/communities")
 public class BoardController {
 
-    private final BoardService boardService;
-
     // todo: 회원별로 접근 가능한 게시판 목록 조회 기능 개발
     @GetMapping()
     public ResponseEntity<?> findAccessibleCodes() {
@@ -29,15 +23,5 @@ public class BoardController {
             accessibleCodeList.add(String.valueOf(boardCode));
         }
         return ResponseEntity.ok().body(accessibleCodeList);
-    }
-
-    @GetMapping("/{code}")
-    public ResponseEntity<?> findPostsByCodeAndCategory(
-            @PathVariable(value = "code") String code,
-            @RequestParam(value = "category", defaultValue = "전체") String category) {
-
-        List<BoardFindPostResponse> postsByCodeAndPostCategory = boardService
-                .findPostsByCodeAndPostCategory(code, category);
-        return ResponseEntity.ok().body(postsByCodeAndPostCategory);
     }
 }
