+++ src/main/java/com/example/solidconnection/community/controller/post/PostController.java
@@ -1,5 +1,6 @@
 package com.example.solidconnection.community.controller.post;
 
+import com.example.solidconnection.community.dto.post.PostListResponse;
 import com.example.solidconnection.custom.resolver.AuthorizedUser;
 import com.example.solidconnection.community.dto.post.PostCreateRequest;
 import com.example.solidconnection.community.dto.post.PostCreateResponse;
@@ -38,6 +39,16 @@ public class PostController {
     private final PostQueryService postQueryService;
     private final PostCommandService postCommandService;
     private final PostLikeService postLikeService;
+
+    @GetMapping("/{code}")
+    public ResponseEntity<?> findPostsByCodeAndCategory(
+            @PathVariable(value = "code") String code,
+            @RequestParam(value = "category", defaultValue = "전체") String category) {
+
+        List<PostListResponse> postsByCodeAndPostCategory = postQueryService
+                .findPostsByCodeAndPostCategory(code, category);
+        return ResponseEntity.ok().body(postsByCodeAndPostCategory);
+    }
 
     @PostMapping(value = "/{code}/posts")
     public ResponseEntity<?> createPost(
