+++ src/main/java/com/example/solidconnection/community/service/post/PostQueryService.java
@@ -1,7 +1,10 @@
 package com.example.solidconnection.community.service.post;
 
+import com.example.solidconnection.community.domain.board.Board;
 import com.example.solidconnection.community.dto.board.PostFindBoardResponse;
 import com.example.solidconnection.community.dto.comment.PostFindCommentResponse;
+import com.example.solidconnection.community.dto.post.PostListResponse;
+import com.example.solidconnection.community.repository.board.BoardRepository;
 import com.example.solidconnection.community.service.comment.CommentService;
 import com.example.solidconnection.custom.exception.CustomException;
 import com.example.solidconnection.community.domain.post.Post;
@@ -13,24 +16,41 @@ import com.example.solidconnection.service.RedisService;
 import com.example.solidconnection.siteuser.domain.SiteUser;
 import com.example.solidconnection.siteuser.dto.PostFindSiteUserResponse;
 import com.example.solidconnection.type.BoardCode;
+import com.example.solidconnection.type.PostCategory;
 import com.example.solidconnection.util.RedisUtils;
 import lombok.RequiredArgsConstructor;
+import org.apache.commons.lang3.EnumUtils;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+import java.util.stream.Collectors;
 
 import static com.example.solidconnection.custom.exception.ErrorCode.INVALID_BOARD_CODE;
+import static com.example.solidconnection.custom.exception.ErrorCode.INVALID_POST_CATEGORY;
 
 @Service
 @RequiredArgsConstructor
 public class PostQueryService {
 
+    private final BoardRepository boardRepository;
     private final PostRepository postRepository;
+    private final PostLikeRepository postLikeRepository;
     private final CommentService commentService;
     private final RedisService redisService;
     private final RedisUtils redisUtils;
-    private final PostLikeRepository postLikeRepository;
+
+    @Transactional(readOnly = true)
+    public List<PostListResponse> findPostsByCodeAndPostCategory(String code, String category) {
+
+        String boardCode = validateCode(code);
+        PostCategory postCategory = validatePostCategory(category);
+
+        Board board = boardRepository.getByCodeUsingEntityGraph(boardCode);
+        List<Post> postList = getPostListByPostCategory(board.getPostList(), postCategory);
+
+        return PostListResponse.from(postList);
+    }
 
     @Transactional(readOnly = true)
     public PostFindResponse findPostById(SiteUser siteUser, String code, Long postId) {
@@ -69,5 +89,21 @@ public class PostQueryService {
     private Boolean getIsLiked(Post post, SiteUser siteUser) {
         return postLikeRepository.findPostLikeByPostAndSiteUser(post, siteUser)
                 .isPresent();
+    }
+
+    private PostCategory validatePostCategory(String category) {
+        if (!EnumUtils.isValidEnum(PostCategory.class, category)) {
+            throw new CustomException(INVALID_POST_CATEGORY);
+        }
+        return PostCategory.valueOf(category);
+    }
+
+    private List<Post> getPostListByPostCategory(List<Post> postList, PostCategory postCategory) {
+        if (postCategory.equals(PostCategory.전체)) {
+            return postList;
+        }
+        return postList.stream()
+                .filter(post -> post.getCategory().equals(postCategory))
+                .collect(Collectors.toList());
     }
 }
